first pass at setting up to run as a gem
