+++ Gemfile
@@ -6,5 +6,8 @@ git_source(:github) {|repo_name| "https://github.com/#{repo_name}" }
 
 # gem "rails"
 
-gem 'paint'
-gem 'word_wrap'
+#gem 'paint'
+#gem 'word_wrap'
+
+# Specify your gem's dependencies in gatherlogs.gemspec
+gemspec
