+++ lib/gatherlogs/version.rb
@@ -1,0 +1,3 @@
+module Gatherlogs
+  VERSION = "0.1.0"
+end
