+++ profiles/automate/controls/basic.rb
