+++ profiles/automate/controls/issues.rb
