+++ profiles/automate/inspec.yml
