+++ profiles/automate2/controls/issues.rb
