+++ profiles/chef-server/controls/issues.rb
