+++ profiles/chef-server/controls/reporting_issues.rb
