+++ profiles/glresources/libraries/common_logs.rb
