+++ profiles/glresources/libraries/installed_packages.rb
