+++ profiles/glresources/libraries/log_analysis.rb
