+++ profiles/glresources/libraries/platform_version.rb
