+++ profiles/glresources/libraries/service_status.rb
